backend/requirements.txt
fastapi==0.111.0
# ASGIサーバー
uvicorn[standard]==0.35.0
# データバリデーション
pydantic==2.7.4
# ORM (データベース操作)
SQLAlchemy==2.0.42
# PostgreSQLデータベースドライバ
psycopg[binary]==3.2.1
# .envファイル読み込み
python-dotenv==1.1.1
# AWS S3
boto3==1.34.0
# ファイルアップロード
python-multipart==0.0.7